CRYPTO-126 Refactor graphql part of the code base
Graphql part of the code base was changed in a way that keeps elements of the similar
nature together e.g. resolvers and schema definitions. Apart from that some minor changes to
names were made in other parts of the project.

This change was neccessary because of unclear structure that was present before.